src/com/zjxjwxk/leetcode/_0300_Longest_Increasing_Subsequence/Solution.java
package com.zjxjwxk.leetcode._0300_Longest_Increasing_Subsequence;

/**
 * 动态规划
 * @author Xinkang Wu
 * @date 2020/11/28 10:52
 */
public class Solution {

    public int lengthOfLIS(int[] nums) {
        if (nums.length == 0) {
            return 0;
        }
        int len = nums.length, ans = 1;
        int[] dp = new int[len];
        dp[0] = 1;
        for (int i = 1; i < len; ++i) {
            dp[i] = 1;
            for (int j = 0; j < i; ++j) {
                if (nums[i] > nums[j]) {
                    dp[i] = Math.max(dp[i], dp[j] + 1);
                }
            }
            ans = Math.max(ans, dp[i]);
        }
        return ans;
    }
}
